ref: update items search controller
