+++ app/controllers/api/v1/items_controller.rb
@@ -1,25 +1,13 @@
 class Api::V1::ItemsController < ApplicationController
     def index 
-        # render json: ItemSerializer.new(Item.all)
         json_response(ItemSerializer.new(Item.all), 200)
     end 
 
      def show 
-        # if Item.exists?(params[:id]) 
-        #     render json: ItemSerializer.new(Item.find(params[:id]))
-        # else 
-        #     render status: 404
-        # end 
         json_response(ItemSerializer.new(Item.find(params[:id])), 200)
     end 
 
     def create
-        # item = Item.new(item_params)
-        # if item.save
-        #     render json: ItemSerializer.new(item), status: 201
-        # else 
-        #     render json: { error: "Item was not created" }, status: 404
-        # end 
         json_response(ItemSerializer.new(Item.create!(item_params)), 201)
     end 
 
@@ -33,7 +21,7 @@ class Api::V1::ItemsController < ApplicationController
             end 
         else 
             render status: 404
-        end 
+        end  
     end 
 
     def destroy 
@@ -42,38 +30,6 @@ class Api::V1::ItemsController < ApplicationController
             item.destroy
         else 
             render status: 404
-        end 
-    end 
-
-    def find 
-        if params[:name] != nil && params[:name] != "" && params[:min_price] == nil && params[:max_price] == nil
-            render json: Item.find_by_name(params[:name])
-        elsif params[:min_price] != nil && params[:min_price].to_f >= 0 && params[:max_price] == nil && params[:name] == nil && params[:min_price] != ''
-            render json: Item.find_by_min_price(params[:min_price])
-        elsif params[:max_price] != nil && params[:max_price].to_f >= 0 && params[:min_price] == nil && params[:name] == nil && params[:max_price] != ''
-            render json: Item.find_by_max_price(params[:max_price])
-        elsif params[:max_price] != nil && params[:min_price] != nil && params[:name] == nil && params[:min_price].to_f < params[:max_price].to_f
-            render json: Item.find_by_price_range(params[:min_price], params[:max_price])
-        elsif params[:min_price].to_f < 0 || params[:max_price].to_f < 0
-            render json: { error: "Number cannot be negative" }, status: 400 
-        else 
-            render status: 400
-        end 
-    end 
-
-     def find_all  
-        if params[:name] != nil && params[:name] != "" && params[:min_price] == nil && params[:max_price] == nil
-            render json: Item.find_all_by_name(params[:name])
-        elsif params[:min_price] != nil && params[:min_price].to_f >= 0 && params[:max_price] == nil && params[:name] == nil && params[:min_price] != ''
-            render json: Item.find_all_by_min_price(params[:min_price])
-        elsif params[:max_price] != nil && params[:max_price].to_f >= 0 && params[:min_price] == nil && params[:name] == nil && params[:max_price] != ''
-            render json: Item.find_all_by_max_price(params[:max_price])
-        elsif params[:max_price] != nil && params[:min_price] != nil && params[:name] == nil && params[:min_price].to_f < params[:max_price].to_f
-            render json: Item.find_all_by_price_range(params[:min_price], params[:max_price])
-        elsif params[:min_price].to_f < 0 || params[:max_price].to_f < 0
-            render json: { error: "Number cannot be negative" }, status: 400 
-        else 
-            render status: 400
         end 
     end
