+++ app/controllers/concerns/exception_handler.rb
@@ -14,8 +14,8 @@ module ExceptionHandler
         #     json_response({ message: e.message }, :not_found)
         # end
 
-         rescue_from ActiveModel::ValidationError do |e|
-            json_response({ message: e.message }, :unprocessable_entity)
-        end
+        #  rescue_from ActiveModel::ValidationError do |e|
+        #     json_response({ message: e.message }, :unprocessable_entity)
+        # end
     end 
 end
