+++ app/controllers/concerns/response.rb
@@ -1,5 +1,5 @@
 module Response 
-    def json_response(object, status = 200)
+    def json_response(object, status = :ok)
         render json: object, status: status
     end 
 end
