+++ config/routes.rb
@@ -2,10 +2,10 @@ Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
   namespace :api do
     namespace :v1 do
-      get '/items/find', to: 'items#find'
-      get '/items/find_all', to: 'items#find_all'
-      get '/merchants/find_all', to: 'merchants#find_all'
-      get '/merchants/find', to: 'merchants#find'
+      get '/items/find', to: 'items/search#find'
+      get '/items/find_all', to: 'items/search#find_all'
+      get '/merchants/find_all', to: 'merchants/search#find_all'
+      get '/merchants/find', to: 'merchants/search#find'
       resources :merchants, only: [:index, :show, :create] do 
         resources :items, controller: :merchant_items, only: [:index]
       end
