+++ spec/requests/api/v1/items/search_spec.rb
@@ -1,0 +1,8 @@
+require 'rails_helper'
+
+describe "Items Search API" do
+    describe 'happy path testing' do 
+
+        
+    end 
+end
